requirements.txt
aiofiles==24.1.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
Telethon==1.37.0
typing_extensions==4.12.2
uvicorn==0.31.0
